+++ VirtualH89/Src/H47Drive.h
@@ -18,7 +18,7 @@ class H47Drive: public DiskDrive
     H47Drive();
     virtual ~H47Drive();
 
-    virtual void getControlInfo(unsigned int pos, bool& hole, bool& trackZero,
+    virtual void getControlInfo(unsigned long pos, bool& hole, bool& trackZero,
                                 bool& writeProtect);
 
     virtual void selectSide(BYTE side);
